awesome_print and tests for WorkWeek#update
awesome_print is very useful for development purposes and I want to be
able to use it outside of the test environment. If there's a less
obstrusive way to achieve that goal, please let me know.

Tests for #update controller method is kind of naive and remains to be
improved as far as the whitelist_attributes private method is involved.