Added current_company on User model
As a result, the view for aggregated staffplans is now filtered through
the current_company of the current_user.

Example: As a user browsing the website for account/company "GoInvo", I
should only see estimates and weekly workloads on the /staffplans page
for people related to that company/account.

TODO: Should write tests for that feature. As in: "as a user, when
browsing the /staffplans page, I shouldn't see anything for users not
working for the same account/company I currently browse the pages
through.