Added invitations
Only admins can send invitations (through the Add Staff page).
Basically adding staff means that the user is created, an invitation is
sent to him or her. When he or she clicks on the invitation in the
email, he or she is redirected to the edit user page and is then
allowed to choose a new password or modify his or her credentials.