Added CRUD stuff for companies
If we're going to have invites by a company administrator, then I
figured we might want to change the administrator or the company name
in case of typo. All the other actions are merely copy pasta from
clients so NBD.

I'll have to add tests for all this if we keep it.