Fix for faulty migration
So what happened is that I later added a name method to the User model
that uses the first_name and last_name to keep
compatibility with older code. Still, at this point in the migrations,
first_name and last_name are not yet defined in the DB but are in the
code since the migration is lagging behind in time.
So yeah basically a Catch-22. I'd be really interested in insight about
how I could have avoided that.