Added roles on memberships
Membership is the name of the relationship between User and Company.

This is where the roles should reside (i.e. a user is an admin for a
particular company, same for contractor and employee).

Also added the assignment of said roles by the company admin in
users#edit and made it so that the first user for a given company is
automatically an admin.