Added financial data and associated permissions
Also added fields in users#edit and #new, projects#new and #edit so that
the financial information can be modified.

TODO: Javascript to restrict the information one can enter upon
selecting the proper employment status (employee?, contractor?,
intern?).

The default values and expected behavior for all actions on users and
projects in that regard need to be clearly defined too.