Added currentCompany to the bootstrap.
We might want to access at the last the id and the name of the
currentCompany in the different views.

As such, I added a reference to the current company of the current user
in the clients list. The variable is accessible as @currentCompany in
the router and can be easily passed along with the request callbacks
used to render the different views.