Fixing proposed hours on project page
Lots of hacks around this stuff. Sharing the component for charts is
actually not a real time-saver.

TODO: Maybe find a better way to discriminate inside chart_totals_view
so that we can get the information about whether or not the project is
proposed.

But really, what we should be doing is that we should bootstrap the
informationb about the assignments.

Either all assignments for user/:id on the staffplan page or the
project/:id assignments on the project page. That way, the code inside
the chart is the same. See with Rob.