Slight refactoring of generic code in chart_totals
Since the view is shared by the projects#show page and the
staffplans#show page, it was not telling the two pages apart in the
code, resulting in the proposed part of the hours not being rendered
properly in the graph.

I thus added a test on the page so that the proposed hours are now
rendered prooperly on both pages.