Added separate controller for memberships
That way we don't mix apples and oranges anymore.

Actions on memberships are done on the MembershipsController and don't
add any clutter to the UsersController or the CompaniesController.i

The creation of a user is thus a two-step process client side now.

1) Create user, display any errors or message.
2) If successful, create the membership.
3) If successful, add it to the user.
4) If not display message.

Each subsequent action on the client-side will have to provide the
proper url and access the proper controllers to avoid overlap in
features.