Cleanup of BB views and models
The membership model client-side now has a proper toJSON method.

We're also not adding a newly created user to the list of users for the
current company since that can result in duplicate keys for
user_id/company_id.

Instead we create the user and later create the associated membership
using the newly created user_id and company_id passed as a parameter.

Also cleaned markup for list of users (unnecessary li's)