Make sure selectedSubset returns a collection
Problem is that all the methods available on collection actually don't
return an actual collection but instead an array of objects.

Because of that you can't simply filter data and expect to be able to
use .get .set and all other model methods right off the bat.

Temporary solution is to create a collection on the fly to pass back to
the caller.