"Proposed" now behaves as intended for new users
When posting to Users::ProjectsController to create a new client and
project, we now return the assignment to the target user in the JSON
response so that it might be added to window._meta.assignments in the
sucess callback.

At this point, when the user tries to switch the proposed bit, the
assignment is there for the detect'ing and it works.