Increases charts' rendering speed
All chart can call a isInViewport() function to know whether or not
they're currently visible in the viewport.

If they are, the chart is updated immediately.

If they are not, the rendering is slightly delayed.

That way, we can have the feeling that the view is updated right away.
We might want to tweak that with a more subtle approach, maybe using
_.after in the following manner:

_.after all visible views have been rendered, render the invisible ones.
That way we don't even have to provide a timeout and everything is done
ASAP.