Juhan will now receive emails on registrations
When the user registers through the login page, Juhan will receive an
email indicating that someone just registered a new company and a new
associated user.

When a user is added through the "Add Staff" functionality on the
staffplans page, Juhan will receive an email indicating that the user
was added to the current company's staffplan.