Bug #40472445 and #40472451
Removed the arrows for navigation in the date range component and
replaced them by key listeners (using keymaster).

Both pages now react to left and right arrow for pagination.

Also improved the calculation of the space the charts should take on the
page so that it's good at all sizes.