Features [#40423729] and [#40423797]
Client names and project names now autocomplete.

Meaning that when you start typing a client's name you get a list of all
the existing clients partially matching that client name.
Same for projects.

Additionally, when the user ends up selecting an existing client, the
list of possible typeaheads for projects is subsequently restricted to
the client's projects.