Features [#40849897] and [#39518145]
Modified company decorator so that it now returns all assignments tied
to projects linked to the current company, disregarding user constraints
(since it now has to return all assignments regardless of whether or not
they are actually tied to a user).

The pages are now refreshed on successful save to the server so that the
TBD user is added to the list of assignees directly.