Bug [#43512327] Redirect to login on sign out
Added a polling mechanism in the staffplan-app file. It polls every 10
second to make sure that there is connectivity and that the credentials
are good on the server. The user gets redirected to the login page if
they're not. We might want to refine the error message or the conditions
leading to the redirect in the future (maybe use that for the other
"similar" feature too).