Feature [#42660023]
Added the rake task to perform the distribution of emails for the
forgetful.

Slightly modified the code handling the checkbox for the "receive
reminder" preference and checked the whole Scheduler Heroku thing.

Seems like the best it can do is daily so we'll have to test the date
inside the task itself.