Feature [#39518213] Custom Profile Picture
Works: Choosing the file, seeing a preview, uploading the picture,
having the picture set as the profile picture, getting a response with
the URL of the avatar created by Paperclip, replacing the avatar by the
new one.

Still missing: Modifying the bootstrap js, styling, S3 integration (we
need credentials for that to happen).